+++ src/dependency/core/agrupation/entrypoint.py
@@ -3,9 +3,9 @@ import time
 from threading import Event
 from typing import Iterable
 from dependency.core.agrupation.plugin import Plugin
-from dependency.core.injection.injectable import Injectable
 from dependency.core.resolution.container import Container
 from dependency.core.resolution.resolver import InjectionResolver
+from dependency.core.resolution.strategy import ResolutionStrategy
 _logger = logging.getLogger("dependency.loader")
 
 class Entrypoint:
@@ -14,25 +14,21 @@ class Entrypoint:
     Attributes:
         init_time (float): Time when the entrypoint was initialized.
     """
-    init_time: float = time.time()
-
     def __init__(self,
         container: Container,
-        plugins: Iterable[type[Plugin]]
+        plugins: Iterable[type[Plugin]],
+        strategy: ResolutionStrategy = ResolutionStrategy()
     ) -> None:
-        providers: list[Injectable] = []
-
-        for plugin in plugins:
-            plugin.resolve_container(container=container)
-            providers.extend(plugin.resolve_providers())
+        init_time: float = time.time()
 
         self.resolver: InjectionResolver = InjectionResolver(
             container=container,
-            providers=providers
         )
-
-        self.resolver.resolve_dependencies()
-        _logger.info(f"Application started in {time.time() - self.init_time} seconds")
+        self.resolver.resolve_dependencies(
+            modules=plugins,
+            strategy=strategy,
+        )
+        _logger.info(f"Application started in {time.time() - init_time} seconds")
 
     def main_loop(self) -> None:
         """Main loop for the application. Waits indefinitely."""