+++ src/dependency/core/resolution/resolver.py
@@ -1,6 +1,8 @@
 from typing import Iterable
 from dependency.core.injection.injectable import Injectable
+from dependency.core.injection.mixin import ContainerMixin
 from dependency.core.resolution.container import Container
+from dependency.core.resolution.registry import Registry
 from dependency.core.resolution.strategy import ResolutionStrategy
 
 class InjectionResolver:
@@ -8,22 +10,64 @@ class InjectionResolver:
     """
     def __init__(self,
         container: Container,
-        providers: Iterable[Injectable],
     ) -> None:
         self.container: Container = container
-        self.providers: list[Injectable] = list(providers)
 
     def resolve_dependencies(self,
-        strategy: type[ResolutionStrategy] = ResolutionStrategy
+        modules: Iterable[type[ContainerMixin]],
+        strategy: ResolutionStrategy = ResolutionStrategy()
+    ) -> list[Injectable]:
+        """Resolve dependencies for a list of modules.
+
+        Args:
+            modules (Iterable[type[ContainerMixin]]): The list of module classes to resolve.
+            strategy (type[ResolutionStrategy]): The resolution strategy to use.
+
+        Returns:
+            list[Injectable]: List of resolved injectables.
+        """
+        Registry.validation()
+        providers = self.resolve_modules(
+            modules=modules
+        )
+        return self.resolve_providers(
+            providers=providers,
+            strategy=strategy
+        )
+
+    def resolve_modules(self,
+        modules: Iterable[type[ContainerMixin]],
+    ) -> list[Injectable]:
+        """Resolve all modules and their dependencies.
+
+        Args:
+            modules (Iterable[type[ContainerMixin]]): The list of module classes to resolve.
+
+        Returns:
+            list[Injectable]: List of resolved injectables from the modules.
+        """
+        providers: list[Injectable] = []
+
+        for module in modules:
+            module.inject_container(container=self.container)
+            providers.extend(module.resolve_providers())
+
+        return providers
+
+    def resolve_providers(self,
+        providers: Iterable[Injectable],
+        strategy: ResolutionStrategy = ResolutionStrategy()
     ) -> list[Injectable]:
         """Resolve all dependencies and initialize them.
 
         Args:
+            providers (Iterable[Injectable]): The list of providers to resolve.
             strategy (type[ResolutionStrategy]): The resolution strategy to use.
 
         Returns:
-            list[Injectable]: List of resolved injectables."""
+            list[Injectable]: List of resolved injectables.
+        """
         return strategy.resolution(
             container=self.container,
-            providers=self.providers,
+            providers=list(providers),
         )