+++ src/dependency/core/resolution/strategy.py
@@ -1,8 +1,14 @@
 import logging
 from pydantic import BaseModel
+from typing import Optional
 from dependency.core.injection.injectable import Injectable
 from dependency.core.resolution.container import Container
 from dependency.core.resolution.errors import raise_resolution_error
+from dependency.core.exceptions import (
+    DeclarationError,
+    InitializationError,
+    CancelInitialization,
+)
 _logger = logging.getLogger("dependency.loader")
 
 class ResolutionConfig(BaseModel):
@@ -13,105 +19,136 @@ class ResolutionConfig(BaseModel):
 class ResolutionStrategy:
     """Defines the strategy for resolving dependencies.
     """
-    config: ResolutionConfig = ResolutionConfig()
+    def __init__(self,
+        config: Optional[ResolutionConfig] = None
+    ) -> None:
+        self.config: ResolutionConfig = config or ResolutionConfig()
 
-    @classmethod
-    def resolution(cls,
-        container: Container,
+    def resolution(self,
         providers: list[Injectable],
+        container: Container,
     ) -> list[Injectable]:
         """Resolve all dependencies and initialize them.
 
         Args:
+            providers (list[Injectable]): List of providers to resolve.
             container (Container): The container to wire the injectables with.
-            providers (list[ProviderInjection]): List of provider injections to resolve.
-            config (ResolutionConfig): Configuration for the resolution strategy.
 
         Returns:
             list[Injectable]: List of resolved injectables.
         """
-        injectables: list[Injectable] = cls.injection(
+        providers = self.expand(
+            providers=providers
+        )
+        self.injection(
             providers=providers,
         )
-        cls.wiring(
+        self.wiring(
+            providers=providers,
             container=container,
-            injectables=injectables,
         )
-        cls.initialize(
-            injectables=injectables
+        self.initialize(
+            providers=providers
         )
-        return injectables
+        return providers
 
-    @classmethod
-    def injection(cls,
+    def expand(self,
         providers: list[Injectable],
     ) -> list[Injectable]:
+        """Expand the list of providers by adding all their imports.
+
+        Args:
+            providers (list[Injectable]): List of providers to expand.
+
+        Returns:
+            list[Injectable]: List of expanded providers.
+        """
+        _logger.info("Expanding dependencies...")
+        unexpanded: set[Injectable] = set(providers.copy())
+        expanded: set[Injectable] = set()
+
+        while unexpanded:
+            provider: Injectable = unexpanded.pop()
+            expanded.add(provider)
+
+            if not provider.partial_resolution:
+                unexpanded.update(filter(lambda i: i not in expanded, provider.imports))
+        return list(expanded)
+
+    def injection(self,
+        providers: list[Injectable],
+    ) -> None:
         """Resolve all injectables in layers.
 
         Args:
             providers (list[Injectable]): List of injectables to resolve.
 
         Returns:
-            list[Injectable]: List of resolved injectables.
+            list[Injectable]: List of unresolved injectables.
         """
         _logger.info("Resolving dependencies...")
-        unresolved: list[Injectable] = providers.copy()
-        resolved: list[Injectable] = []
-        layer_count: int = 0
+        unresolved: set[Injectable] = set(providers.copy())
+        resolved: set[Injectable] = set()
 
         while unresolved:
-            new_layer: set[Injectable] = {
-                provider.inject()
-                for provider in unresolved
-                if provider.import_resolved
-            }
+            layer_resolved: set[Injectable] = set()
+            layer_unresolved: set[Injectable] = set()
 
-            if not new_layer:
+            for provider in unresolved:
+                if provider.check_resolved(providers):
+                    layer_resolved.add(provider)
+                else:
+                    layer_unresolved.add(provider)
+
+            if not layer_resolved:
                 raise_resolution_error(
                     providers=providers,
-                    unresolved=unresolved
+                    unresolved=list(unresolved),
                 )
-            resolved.extend(new_layer)
-            _logger.debug(f"Layer {layer_count}: {new_layer}")
-            layer_count += 1
 
-            for provider in new_layer:
-                unresolved.extend(provider.products)
+            resolved.update(layer_resolved)
+            unresolved = layer_unresolved
 
-            unresolved = [
-                provider
-                for provider in unresolved
-                if provider not in new_layer
-            ]
-        return resolved
-
-    @classmethod
-    def wiring(cls,
+    def wiring(self,
+        providers: list[Injectable],
         container: Container,
-        injectables: list[Injectable],
     ) -> None:
-        """Wire a list of injectables with the given container.
+        """Wire a list of providers with the given container.
 
         Args:
-            container (Container): The container to wire the injectables with.
-            injectables (list[Injectable]): List of injectables to wire.
+            providers (list[Injectable]): List of providers to wire.
+            container (Container): The container to wire the providers with.
         """
         _logger.info("Wiring dependencies...")
-        for injectable in injectables:
-            injectable.wire(container=container)
-        if cls.config.init_container:
+        for provider in providers:
+            container.wire(
+                modules=provider.modules_cls,
+                warn_unresolved=True,
+            )
+        if self.config.init_container:
             container.check_dependencies()
             container.init_resources()
 
-    @classmethod
-    def initialize(cls,
-        injectables: list[Injectable],
+    def initialize(self,
+        providers: list[Injectable],
     ) -> None:
         """Start all implementations by executing their init functions.
 
         Args:
-            injectables (list[Injectable]): List of injectables to start.
+            providers (list[Injectable]): List of providers to start.
         """
         _logger.info("Initializing dependencies...")
-        for injectable in injectables:
-            injectable.init()
+        for provider in providers:
+            if not provider.is_resolved:
+                raise DeclarationError(
+                    f"Injectable {provider} must be resolved before initialization. "
+                    f"Ensure it is declared as a dependency where it is being used"
+                )
+
+            if provider.bootstrap is not None:
+                try:
+                    provider.bootstrap()
+                except CancelInitialization as e:
+                    _logger.warning(f"Injectable {provider} initialization skipped (cancelled by user): {e}")
+                except Exception as e:
+                    raise InitializationError(f"Injectable {provider} initialization failed") from e