+++ tests/core/test_declaration.py
@@ -2,6 +2,7 @@ from abc import ABC, abstractmethod
 from dependency_injector import providers
 from dependency.core.agrupation import Module, module
 from dependency.core.declaration import Component, component, instance
+from dependency.core.injection import Injectable
 from dependency.core.resolution import Container
 
 @module()
@@ -26,12 +27,12 @@ class TInstance(TComponent):
 def test_declaration() -> None:
     container = Container()
     TModule.inject_container(container)
-    for provider in TModule.injection.resolve_providers():
-        provider.inject()
+    injectables: list[Injectable] = list(TModule.resolve_providers())
+    for provider in injectables:
+        assert provider.check_resolved(injectables)
 
-    assert TComponent.injection.injectable.interface_cls == TComponent
-    assert TComponent.injection.injectable.implementation == TInstance
-    assert TInstance.injection.injectable.is_resolved
+    assert TComponent.injectable.interface_cls == TComponent
+    assert TComponent.injectable.implementation == TInstance
 
     component: TComponent = TComponent.provide()
     assert isinstance(component, TComponent)