+++ tests/core/test_interfaces.py
@@ -3,7 +3,7 @@ from dependency_injector import providers
 from dependency_injector.wiring import inject
 from dependency.core.agrupation import Module, module
 from dependency.core.declaration import Component, component
-from dependency.core.injection import LazyProvide
+from dependency.core.injection import Injectable, LazyProvide
 from dependency.core.resolution import Container
 
 @module()
@@ -48,17 +48,18 @@ def test_interfaces() -> None:
     container = Container()
     TModule.inject_container(container)
 
-    injectables = list(TModule.resolve_providers())
+    injectables: list[Injectable] = list(TModule.resolve_providers())
     for injectable in injectables:
-        injectable.inject()
-    for injectable in injectables:
-        injectable.wire(container)
+        injectable.check_resolved(injectables)
+    assert TProduct1.injectable.check_resolved(injectables)
+    assert TProduct2.injectable.check_resolved(injectables)
 
+    for injectable in injectables:
+        container.wire(injectable.modules_cls)
     product1: TProduct1 = TProduct1.provide()
     product2: TProduct2 = TProduct2.provide()
 
     assert isinstance(product1, TProduct1)
     assert isinstance(product2, TProduct2)
-
     assert product1.method() == "Hello, World!"
     assert product2.method() == "Hello, World!"