+++ tests/core/test_products.py
@@ -15,14 +15,12 @@ class TPlugin(Plugin):
 class TComponent1(Component):
     pass
 
-@component(
-)
+@component()
 class TComponent2(Component):
     pass
 
 @component(
     imports=[
-        TComponent1,
         TComponent2,
     ],
     provider=providers.Factory,
@@ -31,23 +29,28 @@ class TProduct1(Component):
     pass
 
 @instance(
-    products=[
-        TProduct1
+    imports=[
+        TProduct1,
     ],
 )
 class TInstance1(TComponent1):
     pass
 
 def test_products() -> None:
+    strategy: ResolutionStrategy = ResolutionStrategy()
     container = Container()
+
     TPlugin.resolve_container(container)
     injectables = list(TPlugin.resolve_providers())
+    assert injectables == [TComponent1.injectable]
 
     with pytest.raises(ResolutionError):
-        ResolutionStrategy.injection(injectables)
+        expanded = strategy.expand(injectables)
+        strategy.injection(expanded)
 
-    TProduct1.injection.injectable.partial_resolution = True
-    injectables = ResolutionStrategy.injection(injectables)
+    TProduct1.injectable.partial_resolution = True
+    expanded = strategy.expand(injectables)
+    strategy.injection(expanded)
 
-    assert TComponent1.injection.injectable in injectables
-    assert TProduct1.injection.injectable in injectables
+    assert TComponent1.injectable.is_resolved
+    assert TProduct1.injectable.is_resolved