+++ tests/core/test_resolution.py
@@ -1,4 +1,5 @@
 import pytest
+from dependency_injector import providers
 from dependency.core.agrupation import Plugin, PluginMeta, module
 from dependency.core.declaration import Component, component, instance, providers
 from dependency.core.resolution import Container, InjectionResolver
@@ -23,14 +24,15 @@ class TComponent2(Component):
     pass
 
 @component(
-    imports=[TComponent1],
     provider=providers.Factory,
 )
 class TProduct1(Component):
     pass
 
 @instance(
-    products=[TProduct1],
+    imports=[
+        TProduct1,
+    ],
     bootstrap=True,
 )
 class TInstance1(TComponent1):
@@ -46,15 +48,15 @@ class TInstance2(TComponent2):
         BOOTSTRAPED.append("TInstance2")
         raise CancelInitialization("Failed to initialize TInstance2")
 
-def test_exceptions() -> None:
+def test_resolution() -> None:
     container = Container.from_json("example/config.json")
     injectables = TPlugin.resolve_providers()
     assert "TInstance1" not in BOOTSTRAPED
 
-    loader = InjectionResolver(container, injectables)
+    loader = InjectionResolver(container)
     assert "TInstance1" not in BOOTSTRAPED
 
-    loader.resolve_dependencies()
+    loader.resolve_providers(injectables)
     assert "TInstance1" in BOOTSTRAPED
     assert "TInstance2" in BOOTSTRAPED